refactor(api): move managedai out of internal/ so plugins can reuse it
The managed-inference preset (managedai.Apply / Enabled) is the single
source of truth for how a project's LLM/blurb/embedding config is
canonicalized onto the gateway in managed mode. Enterprise plugins that
create projects on their own path (e.g. pack-gen drafts) need to apply
the same preset, but Go's internal/ rule blocked them from importing
services/api/internal/managedai.

Relocate the package to services/api/managedai (name and logic
unchanged) and update the in-tree importers. No behavior change.
